requirements.txt
black==22.3.0
boto3==1.24.13
botocore==1.27.13
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
fuzzywuzzy==0.18.0
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
googlemaps==4.6.0
httplib2==0.20.4
idna==3.3
jmespath==1.0.1
mypy==0.961
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
protobuf==4.21.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfiglet==0.8.post1
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
python-Levenshtein==0.12.2
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
